requirements.txt
torch
torchvision
matplotlib
flask
Pillow
